Agregado de primera vista para avion simple el mapa ahora posee la capacidad de agregar aviones al gameloop (no eliminar todavia). Hay un bug en el avance de los aviones que hacen que vuelvan sobre sus pasos, lo voy a arrerglar en la proxima subida. La clase nivel esta rediseñada y a medida que arregle los bugs que quedan la voy a ir incorporando en el juego.

+++ TrabajoPractico2/src/modelo/Aviones/EstrategiaAvionComputarizado.java
@@ -6,7 +6,7 @@ import modelo.Utilitarios.Vector;
 
 public class EstrategiaAvionComputarizado implements EstrategiaAvion {
 
-	private static double tamaño = 10;
+	private static int tamaño = 10;
 	
 	public Vector avanzar(Vector posicion, double velocidad, Trayectoria trayectoriaDeVuelo)
 	{
@@ -43,7 +43,7 @@ public class EstrategiaAvionComputarizado implements EstrategiaAvion {
 	}
 
 
-	public double tamaño() 
+	public int tamaño() 
 	{
 		return this.tamaño;
 	}
